backend/requirements.txt
accelerate
certifi
charset-normalizer
colorama
contourpy
cycler
fastapi
filelock
fonttools
fsspec
huggingface-hub
idna
Jinja2
kiwisolver
MarkupSafe
matplotlib
mpmath
networkx
numpy
opencv-python
packaging
pillow
pillow_heif
psutil
pyparsing
python-dateutil
PyYAML
regex
requests
safetensors
scipy
setuptools
six
sympy
tokenizers
tqdm
transformers
typing_extensions
urllib3
uvicorn[standard]
